fix types (one more love)

+++ src/msspi.cpp
@@ -1897,7 +1897,7 @@ static CERT_USAGE is_cert_usage( PCCERT_CONTEXT pcert, const char * oid )
         if( CertGetEnhancedKeyUsage( pcert, 0, ekuList, &ekuLength ) )
         {
             if( ekuList->cUsageIdentifier == 0 )
-                return ( GetLastError() == CRYPT_E_NOT_FOUND ) ? CERT_USAGE::EVERYTHING : CERT_USAGE::NOT_FOUND;
+                return ( GetLastError() == (DWORD)CRYPT_E_NOT_FOUND ) ? CERT_USAGE::EVERYTHING : CERT_USAGE::NOT_FOUND;
             for( DWORD i = 0; i < ekuList->cUsageIdentifier; i++ )
                 if( 0 == strcmp( ekuList->rgpszUsageIdentifier[i], oid ) )
                     return CERT_USAGE::FOUND;
